+++ src/commands/perform-tasks.ts
@@ -1,11 +1,12 @@
 import { readFileSync } from 'fs';
 import { Command } from 'commander';
+import minimatch from 'minimatch';
 import { BaseCliCommand, ICommandArgs } from './base-command';
 import { UpdateOrganizationCommand } from './update-organization';
 import { BuildTaskProvider } from '~build-tasks/build-task-provider';
 import { ITrackedTask, PersistedState } from '~state/persisted-state';
 import { Validator } from '~parser/validator';
-import { BuildConfiguration } from '~build-tasks/build-configuration';
+import { BuildConfiguration, IBuildTask } from '~build-tasks/build-configuration';
 import { BuildRunner } from '~build-tasks/build-runner';
 import { ConsoleUtil } from '~util/console-util';
 import { S3StorageProvider } from '~state/storage-provider';
@@ -44,6 +45,7 @@ export class PerformTasksCommand extends BaseCliCommand<IPerformTasksCommandArgs
         command.option('--organization-state-bucket-name [organization-state-bucket-name]', 'name of the bucket that contains the read-only organization state');
         command.option('--debug-templating [debug-templating]', 'when set to true the output of text templating processes will be stored on disk', false);
         command.option('--templating-context-file [templating-context-file]', 'json file used as context for nunjuck text templating of organization and tasks file');
+        command.option('--match [match]', 'glob pattern used to define/filter which tasks to run.');
         command.option('--large-template-bucket-name [large-template-bucket-name]', 'bucket used when uploading large templates. default is to create a bucket just-in-time in the target account');
 
         command.option('--skip-storing-state', 'when set, the state will not be stored');
@@ -71,6 +73,13 @@ export class PerformTasksCommand extends BaseCliCommand<IPerformTasksCommandArgs
         const tasks = config.enumBuildTasks(command);
         ConsoleUtil.state = state;
 
+        if (command.match) {
+            const skippedTasks = this.skipNonMatchingLeafTasks(tasks, command.match, '');
+            if (skippedTasks === tasks.length) {
+                ConsoleUtil.LogWarning(`--match parameter glob '${command.match}' did not match any tasks. Use --verbose to see the tasks it did not match`);
+            }
+        }
+
         state.performUpdateToVersion2IfNeeded();
         UpdateOrganizationCommand.ResetHasRan();
 
@@ -90,6 +99,36 @@ export class PerformTasksCommand extends BaseCliCommand<IPerformTasksCommandArgs
 
         await state.save();
 
+    }
+
+    private skipNonMatchingLeafTasks(tasks: IBuildTask[], taskMatcher: string, tasksPrefix: string): number {
+        let skippedTasks = 0;
+        for (const task of tasks) {
+
+            const isLeafTask = task.childTasks.length === 0;
+            const taskFullName = `${tasksPrefix}${task.name}`;
+
+            if (isLeafTask) {
+                const isMatching = task.name === taskMatcher || minimatch(taskFullName, taskMatcher);
+                task.skip = isMatching ? false : true;
+            } else {
+                const skippedChildTasks = this.skipNonMatchingLeafTasks(task.childTasks, taskMatcher, `${taskFullName}/`);
+                const isAllSkipped = task.childTasks.length === skippedChildTasks;
+                task.skip = isAllSkipped ? true : false;
+            }
+
+            if (task.skip) {
+                skippedTasks = skippedTasks + 1;
+            }
+
+            if (isLeafTask && task.skip !== true) {
+                ConsoleUtil.LogInfo(`${taskFullName} matched the '${taskMatcher}' globPattern`);
+            } else {
+                ConsoleUtil.LogDebug(`${taskFullName} did not match the '${taskMatcher}' globPattern`);
+            }
+
+        }
+        return skippedTasks;
     }
 
     public static async PublishChangedOrganizationFileIfChanged(command: IPerformTasksCommandArgs, state: PersistedState): Promise<void> {
@@ -129,4 +168,5 @@ export interface IPerformTasksCommandArgs extends ICommandArgs {
     forceDeploy?: boolean;
     organizationObject?: any;
     skipStoringState?: true;
+    match?: string;
 }